+++ src/refactorings/extract-to-instance-property/index.ts
@@ -1,0 +1,19 @@
+import { RefactoringWithActionProviderConfig } from "../../refactorings";
+import {
+  createVisitor,
+  extractToInstanceProperty
+} from "./extract-to-instance-property";
+
+const config: RefactoringWithActionProviderConfig = {
+  command: {
+    key: "extractToInstanceProperty",
+    operation: extractToInstanceProperty,
+    title: "Extract to Instance Property"
+  },
+  actionProvider: {
+    message: "Extract to instance property",
+    createVisitor
+  }
+};
+
+export default config;